requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cpf==0.1.0
django-recaptcha==3.0.0
mysqlclient==2.1.1
Pillow==9.4.0
python-decouple==3.8
sqlparse==0.4.3
tzdata==2022.7
